Remove reference to remixing
Suggest removing 3.1 altogether, "You may use and remix the Community marks on the Community projects as you please." If the "project" is something distributable and community members are distributing it in modified form (like software), the last thing you want is for them to use a modification of the trademark for their own iteration of their distribution, it dilutes the brand value and expressly permitting it is the definition of naked licensing. I'm not sure in what context that provision makes any sense except for a Wikipedia-type of product.